add useDataFetch hook
It just makes the component a little easier to read, and the data fetch may need to be performed in multiple places later.
